:bug: fix(editor): Make Callout take full width
